backend/requirements.txt
aiosqlite==0.21.0
aistudio-sdk==0.3.8
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bce-python-sdk==0.9.46
bcrypt==4.3.0
boto3==1.40.42
botocore==1.40.42
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorlog==6.9.0
cryptography==46.0.1
cssselect==1.3.0
cssutils==2.11.1
deptry==0.23.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
enum-compat==0.0.3
et_xmlfile==2.0.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
filelock==3.19.1
fsspec==2025.9.0
ftfy==6.3.1
future==1.0.0
google-api-core==2.25.1
google-auth==2.41.0
google-cloud-documentai==3.6.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.35.1
idna==3.10
imageio==2.37.0
imagesize==1.4.1
intuit-oauth==1.2.6
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
lazy_loader==0.4
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
modelscope==1.30.0
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
oauthlib==3.3.1
openai==2.3.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
passlib==1.7.4
pdf2image==1.17.0
pillow==11.3.0
premailer==3.10.0
prettytable==3.16.0
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requirements-parser==0.13.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
s3transfer==0.14.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
sentry-sdk==2.39.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.4
tiktoken==0.11.0
tokenizers==0.22.1
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
websockets==15.0.1
